refactor: moving `Auth` types
